requirements.txt
numpy==2.4.2
PyYAML==6.0.3
matplotlib==3.10.8
jsonschema==4.25.1
